+++ src/microkanren/core.py
@@ -10,12 +10,12 @@ efficient access
 
 from collections.abc import Callable, Generator
 from dataclasses import dataclass
-from functools import reduce, update_wrapper, wraps
+from functools import reduce, wraps
 from typing import Any, Optional, Protocol, TypeAlias, TypeVar
 
+import immutables
 from fastcons import cons, nil
-from pyrsistent import PClass, field, pmap
-from pyrsistent.typing import PMap
+from pyrsistent import PClass, field
 
 from microkanren.utils import identity
 
@@ -38,19 +38,19 @@ class ReifiedVar:
 Value: TypeAlias = (
     Var | int | str | bool | tuple["Value", ...] | list["Value"] | cons | nil
 )
-Substitution: TypeAlias = PMap[Var, Value]
+Substitution: TypeAlias = immutables.Map[Var, Value]
 NeqStore: TypeAlias = list[list[tuple[Var, Value]]]
-DomainStore: TypeAlias = PMap[Var, set[int]]
+DomainStore: TypeAlias = immutables.Map[Var, set[int]]
 ConstraintFunction: TypeAlias = Callable[["State"], Optional["State"]]
 ConstraintStore: TypeAlias = list["Constraint"]
 
 
 def empty_sub() -> Substitution:
-    return pmap()
+    return immutables.Map()
 
 
 def empty_domain_store() -> DomainStore:
-    return pmap()
+    return immutables.Map()
 
 
 def empty_constraint_store() -> ConstraintStore:
@@ -81,10 +81,10 @@ class ConstraintProto(Protocol):
         ...
 
 
-@dataclass(slots=True)
+@dataclass(slots=True, frozen=True)
 class Constraint:
     func: ConstraintProto
-    operands: list[Value]
+    operands: tuple[Value]
 
     def __call__(self, state: State) -> State | None:
         return self.func(*self.operands)(state)
@@ -105,7 +105,6 @@ class GoalConstructorProto(Protocol):
 
 class Goal:
     def __init__(self, goal: GoalProto):
-        update_wrapper(self, goal)
         self.goal = goal
 
     def __call__(self, state: State) -> Stream:
@@ -251,8 +250,8 @@ def delay(g: GoalProto) -> GoalProto:
 
 def disj(g: GoalProto, *goals: GoalProto) -> GoalProto:
     if goals == ():
-        return delay(g)
-    return reduce(_disj, (delay(goal) for goal in goals), delay(g))
+        return g
+    return reduce(_disj, (goal for goal in goals), g)
 
 
 def _disj(g1: GoalProto, g2: GoalProto) -> GoalProto:
@@ -264,8 +263,8 @@ def _disj(g1: GoalProto, g2: GoalProto) -> GoalProto:
 
 def conj(g: GoalProto, *goals: GoalProto) -> GoalProto:
     if goals == ():
-        return delay(g)
-    return reduce(_conj, (delay(goal) for goal in goals), delay(g))
+        return g
+    return reduce(_conj, (goal for goal in goals), g)
 
 
 def _conj(g1: GoalProto, g2: GoalProto) -> GoalProto:
@@ -308,6 +307,26 @@ def unify_all(
             return None
 
 
+def pairs(xs):
+    _xs = iter(xs)
+    while True:
+        try:
+            a = next(_xs)
+        except StopIteration:
+            break
+        try:
+            b = next(_xs)
+            yield (a, b)
+        except StopIteration:
+            raise ValueError("got sequence with uneven length")
+
+
+def unpairs(xs):
+    for a, b in xs:
+        yield a
+        yield b
+
+
 def maybe_unify(
     pair: tuple[Value, Value], sub: Substitution | None
 ) -> Substitution | None:
@@ -325,23 +344,22 @@ def flip(f):
     return _flipped
 
 
-def neq(*pairs) -> GoalProto:
-    return goal_from_constraint(neqc(pairs))
+def neq(u, v, /, *rest) -> GoalProto:
+    return goal_from_constraint(neqc(u, v, *rest))
 
 
-def neqc(pairs: tuple[tuple[Value, Value], ...]) -> ConstraintFunction:
+def neqc(u, v, *rest) -> ConstraintFunction:
     def _neqc(state: State) -> State | None:
-        (u, v), *rest = pairs
-        new_sub = reduce(flip(maybe_unify), rest, unify(u, v, state.sub))
+        new_sub = reduce(flip(maybe_unify), pairs(rest), unify(u, v, state.sub))
         if new_sub is None:
             return state
         elif new_sub == state.sub:
             return None
         prefix = get_sub_prefix(new_sub, state.sub)
-        remaining_pairs = list(prefix.items())
+        remaining_pairs = tuple(prefix.items())
         return state.set(
             constraints=extend_constraint_store(
-                Constraint(neqc, [remaining_pairs]), state.constraints
+                Constraint(neqc, tuple(unpairs(remaining_pairs))), state.constraints
             )
         )
 
@@ -440,7 +458,11 @@ def map_sum(goal_constructor: Callable[[A], GoalProto], xs: list[A]) -> GoalProt
 
 
 def get_sub_prefix(new_sub: Substitution, old_sub: Substitution) -> Substitution:
-    return pmap({k: v for k, v in new_sub.items() if k not in old_sub})
+    mutation = new_sub.mutate()
+    for k in new_sub:
+        if k in old_sub:
+            del mutation[k]
+    return mutation.finish()
 
 
 def fresh(fp: Callable) -> GoalProto: